rename docker-compose.cjs to mock-app-config.cjs
