+++ test/app/database-changelog.spec.js
@@ -1,6 +1,7 @@
 const path = require('path');
 const fse = require('fs-extra');
 const helpers = require('yeoman-test');
+const { createMockedConfig } = require('../support/mock-config.cjs');
 
 const { SERVER_MAIN_RES_DIR } = require('../../generators/generator-constants');
 
@@ -12,7 +13,7 @@ describe('jhipster:app database changelogs', () => {
         helpers
           .create(require.resolve('../../generators/app'))
           .doInDir(dir => {
-            fse.copySync(path.join(__dirname, '../templates/compose/05-cassandra'), dir);
+            createMockedConfig('05-cassandra', dir);
             fse.copySync(path.join(__dirname, '../templates/.jhipster/Simple.json'), path.join(dir, '.jhipster/Foo.json'));
           })
           .withOptions({ withEntities: true, force: true, skipClient: true })
