+++ test/entity/single-entity.spec.js
@@ -3,6 +3,7 @@ const fse = require('fs-extra');
 const helpers = require('yeoman-test');
 
 const { SERVER_MAIN_RES_DIR, SERVER_MAIN_SRC_DIR, CLIENT_MAIN_SRC_DIR } = require('../../generators/generator-constants');
+const { createMockedConfig } = require('../support/mock-config.cjs');
 
 const DEFAULT_TEST_OPTIONS = { fromCli: true, skipInstall: true, skipChecks: true, skipPrettier: true };
 
@@ -51,7 +52,7 @@ describe('jhipster:entity --single-entity', () => {
         helpers
           .create(require.resolve('../../generators/entity'))
           .doInDir(dir => {
-            fse.copySync(path.join(__dirname, '../templates/compose/05-cassandra'), dir);
+            createMockedConfig('05-cassandra', dir);
             fse.copySync(path.join(__dirname, '../templates/.jhipster/Simple.json'), path.join(dir, '.jhipster/Foo.json'));
             fse.copySync(path.join(__dirname, '../templates/.jhipster/Simple2.json'), path.join(dir, '.jhipster/Bar.json'));
           })
